# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRM Platform API",
        "description": "API documentation for HRM Platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/leaves/types": {
            "get": {
                "tags": [
                    "Leaves"
                ],
                "summary": "Get all leave types",
                "operationId": "78ea2d82b3dc761e16d5e6342863ab48",
                "responses": {
                    "200": {
                        "description": "Leave types list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/leaves": {
            "get": {
                "tags": [
                    "Leaves"
                ],
                "summary": "Get leave requests",
                "operationId": "cf3dbd0161e5e14d0e14c12e8e928fe3",
                "responses": {
                    "200": {
                        "description": "Leave requests list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Leaves"
                ],
                "summary": "Create a leave request",
                "operationId": "dcca8df053754c49c6d66bfee9cee011",
                "responses": {
                    "200": {
                        "description": "Leave request created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/leaves/{id}": {
            "put": {
                "tags": [
                    "Leaves"
                ],
                "summary": "Update/Approve/Reject leave request",
                "operationId": "b1b46a5ebe5a311805741ae96d8f6056",
                "responses": {
                    "200": {
                        "description": "Leave request updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/payroll/periods": {
            "get": {
                "tags": [
                    "Payroll"
                ],
                "summary": "Get all payroll periods",
                "operationId": "dfc959b9616bddbf2d70e0fffd9ca940",
                "responses": {
                    "200": {
                        "description": "Payroll periods list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Payroll"
                ],
                "summary": "Create a new payroll period",
                "operationId": "93e36f64d1a2769f71e87cdfa525e756",
                "responses": {
                    "200": {
                        "description": "Payroll period created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/payroll/records": {
            "get": {
                "tags": [
                    "Payroll"
                ],
                "summary": "Get payroll records",
                "operationId": "d50a096a940e53afcd4f50edabccf6f5",
                "responses": {
                    "200": {
                        "description": "Payroll records list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/payroll/calculate": {
            "post": {
                "tags": [
                    "Payroll"
                ],
                "summary": "Calculate payroll for a period",
                "operationId": "6c6393ce727e8f8096a36f0bd14449f5",
                "responses": {
                    "200": {
                        "description": "Payroll calculated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/payroll/my-payslips": {
            "get": {
                "tags": [
                    "Payroll"
                ],
                "summary": "Get current user payslips",
                "operationId": "ff788b471f88268e7f6d6d0d17846d55",
                "responses": {
                    "200": {
                        "description": "User payslips"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/payroll/periods/{id}/approve": {
            "put": {
                "tags": [
                    "Payroll"
                ],
                "summary": "Approve a payroll period",
                "operationId": "6de4f9be9e04838332d86ceb809de3fd",
                "responses": {
                    "200": {
                        "description": "Payroll period approved"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/payroll/records/{id}/pdf": {
            "get": {
                "tags": [
                    "Payroll"
                ],
                "summary": "Download a payslip as PDF",
                "operationId": "45def8fcbb7f7415fc700e22aa66041f",
                "responses": {
                    "200": {
                        "description": "Payslip PDF"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Payslip not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}